build: 更新 C++ 标准和异常处理
- 添加 -DLIBUUID_CPP20_OR_GREATER 编译标志
- 全局启用 C++ 异常处理
- 调整 Android 平台的异常处理位置
- 更新 sqlite_db 文件中的数据类型包含
- 修正 bind_blob 函数参数类型为 size_t

+++ src/core/components/sqlite_db.cpp
@@ -2,8 +2,9 @@
 #include <stdexcept>
 #include <string>
 #include <vector>
+#include <cstddef> 
 
-#include "godot-sqlist/src/sqlite/sqlite3.h"
+#include "gdsqlite/sqlite/sqlite3.h"
 #include "sqlite_db.h"
 
 #include <godot_cpp/classes/file_access.hpp>
@@ -128,7 +129,7 @@ void SQLiteDB::Stmt::bind_double(int idx, double val) {
 	sqlite3_bind_double(stmt_, idx, val);
 }
 
-void SQLiteDB::Stmt::bind_blob(int idx, const void *val, uint size) {
+void SQLiteDB::Stmt::bind_blob(int idx, const void *val, size_t size) {
 	if (!stmt_) {
 		return;
 	}
